switch to using slate (#1)
switches to use slate for rendering the editor. it handles dealing with the contenteditable stuff, and has a data model/API that seems flexible and easy enough to use. it didn't seem possible to simply replace the global state with slate's state for a number of reasons, so instead a hybrid approach is used: the app will manage the global state and extract only the viewable nodes out and pass them to slate. changes are managed locally, but are published back to the global state whenever the node changes (push or pop)

some misc changes along the way:
- split up index.tsx as it gets a little large
- use nanoid for generating IDs for simplicity (avoid having to pipe the state through everywhere). the larger IDs makes things harder to read when debugging but so far its' fine
- use immer for managing state which also avoids the need to have nested reducers